rename pass classes
